Reorganize and break out circlink into multiple files
